+++ src/browser.ts
@@ -2,7 +2,7 @@ import { version } from './utils/version.js';
 import { createEnv, getEnv } from './env/browser/index.js';
 import getHost from './host/in-page/index.js';
 import createSandbox from './sandbox/browser/index.js';
-import { createPublisher } from './publisher/browser/index.js';
+import { createPublisher, connectPublisherWs } from './publisher/browser/index.js';
 import { getSubscriber, getSelfSubscriber } from './subscriber/browser/index.js';
 import { scriptFromFile } from './sdk/browser/scriptFromFile.js';
 
@@ -11,6 +11,7 @@ export {
     getEnv,
     getHost,
     createPublisher,
+    connectPublisherWs,
     getSubscriber,
     getSelfSubscriber,
     scriptFromFile,