+++ src/publisher/index.ts
@@ -1,4 +1,8 @@
-import { establishWsConnection } from './transport-ws.js';
-import { createPublisherFactory } from './factory.js';
+import { getPublisher, connect } from './factory.js';
+import { createNodeWsTransport, fetchWsSettings } from './transport-ws.js';
 
-export const createPublisher = createPublisherFactory(establishWsConnection);
+export const createPublisher = getPublisher;
+
+export function connectPublisherWs(uri?: string) {
+    connect(createNodeWsTransport, fetchWsSettings, uri);
+}