+++ src/publisher/TransportPublisher.ts
@@ -1,30 +1,37 @@
 import { Publisher } from '../classes/Publisher.js';
-import { GetRemoteUIHandler, Options, RequestClientHandler } from './types.js';
+import {
+    GetRemoteUIHandler,
+    Options,
+    GetRemoteUIInternalResult,
+    GetRemoteUISettings,
+} from './types.js';
 
 export class TransportPublisher extends Publisher {
-    getRemoteUI: (settings: unknown, callback: RequestClientHandler) => void;
+    getRemoteUI: (settings: GetRemoteUISettings) => Promise<GetRemoteUIInternalResult>;
     options: Options;
 
     constructor(id: string, getRemoteUI: GetRemoteUIHandler, options?: Options) {
         super(id);
 
         this.options = options || {};
-        this.getRemoteUI = (settings, callback) => {
-            getRemoteUI(
-                settings,
-                (error: string | null, type: 'url' | 'script', content: string) => {
-                    let response: Record<string, string> | string | null = content;
-
-                    if (!error && type === 'script') {
-                        // send with user script rempl source too
-                        response = {
-                            'publisher-ui.js': content,
+        this.getRemoteUI = (settings) => {
+            try {
+                return Promise.resolve(getRemoteUI(settings)).then((result) => {
+                    if (result.type === 'script') {
+                        // TODO: value should be a string
+                        return {
+                            type: 'script',
+                            value: {
+                                'publisher-ui.js': result.value,
+                            },
                         };
                     }
 
-                    callback(error, type, response);
-                }
-            );
+                    return result;
+                });
+            } catch (e: any) {
+                return Promise.reject(e);
+            }
         };
     }
 }