+++ src/publisher/types.ts
@@ -1,12 +1,9 @@
-export type RequestClientHandler = (
-    error: string | null,
-    type: string,
-    response: Record<string, string> | string | null
-) => void;
-export type GetRemoteUIHandler = (
-    settings: unknown,
-    callback: (error: string | null, type: 'url' | 'script', content: string) => void
-) => void;
+export {
+    GetRemoteUISettings,
+    GetRemoteUIHandler,
+    GetRemoteUIInternalResult,
+} from '../transport/types.js';
+
 export type Options = { ws?: string };
 export type WsSettings = {
     explicit: string | undefined;