+++ src/sdk/scriptFromFile.ts
@@ -1,9 +1,9 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import { fileURLToPath } from 'url';
-import { GetRemoteUIFn } from '../transport/event.js';
+import { GetRemoteUIHandler } from '../transport/types.js';
 
-export function scriptFromFile(filename: string, includeRempl = false): GetRemoteUIFn {
+export function scriptFromFile(filename: string, includeRempl = false): GetRemoteUIHandler {
     const remplPath = path.join(
         path.dirname(fileURLToPath(import.meta.url)),
         '../../dist/rempl.js'
@@ -11,18 +11,16 @@ export function scriptFromFile(filename: string, includeRempl = false): GetRemot
     let cache: string | null = null;
 
     // TODO: take in account settings.accept setting
-    return function (settings, callback) {
-        if (settings.dev || cache === null) {
-            fs.readFile(filename, 'utf-8', function (err, content) {
-                if (err) {
-                    return callback(err);
-                }
-
-                cache = (includeRempl ? fs.readFileSync(remplPath, 'utf8') : '') + content;
-                callback(null, 'script', content);
-            });
-        } else {
-            callback(null, 'script', cache);
+    return function (settings) {
+        if (!settings.dev && cache !== null) {
+            return { type: 'script', value: cache };
         }
+
+        return fs.promises.readFile(filename, 'utf-8').then((content) => {
+            return {
+                type: 'script',
+                value: (cache = (includeRempl ? fs.readFileSync(remplPath, 'utf8') : '') + content),
+            };
+        });
     };
 }